pyproject.toml
[build-system]
requires = ["uv_build >= 0.9.18, <0.11.0"]
build-backend = "uv_build"

[project]
name = "prompt-engineering-proxy"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Frederic Morin", email = "frederic.morin.8@gmail.com"},
]

dependencies = [
    "uvicorn",
    "fastapi",
    "pydantic-settings",
    "httpx",
    "aiosqlite",
    "redis[hiredis]",
    "sse-starlette",
    "python-ulid",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx", "ty", "ruff", "uv", "ipdb"]

[tool.ruff]
line-length = 120
